Restore JSONEventLayout class
For people migrating from 1.1, the removal of JSONEventLayout may be
painful as the log4j configuration needs to be updated at the same time
as the package which can be difficult for long-running processes. The
reintroduction of JSONEventLayout as an alias for JSONEventLayoutV0 will
help a smooth migration.

A warning message is added when using JSONEventLayout and the
documentation stresses that its use is discouraged.

Related discussion may be found in #21 and #23.